requirements.txt
fastapi
uvicorn
requests
google-generativeai
python-dotenv
rich
pydantic
